separated reference tab into own component
