+++ resources/js/components/MainView.vue
@@ -61,82 +61,7 @@
                 <keep-alive>
                     <component :is="state.tabComponent" />
                 </keep-alive>
-                <div
-                    v-show="isTab('references') && !!state.entity.id"
-                    class="h-100 overflow-y-auto"
-                >
-                    <p
-                        v-if="!state.hasReferences"
-                        class="alert alert-info"
-                    >
-                        {{ t('main.entity.references.empty') }}
-                    </p>
-                    <div
-                        v-if="state.hasEntityReferences"
-                        class="reference-group mb-1"
-                    >
-                        <h5 class="mb-2 fw-medium">
-                            {{ t('main.entity.references.general') }}
-                        </h5>
-                        <div class="list-group w-90">
-                            <div
-                                v-for="(reference, i) in state.entityReferences"
-                                :key="i"
-                                class="list-group-item pt-0"
-                            >
-                                <header class="text-end">
-                                    <span class="text-muted fw-light small">
-                                        {{ date(reference.updated_at) }}
-                                    </span>
-                                </header>
-                                <Quotation :value="reference" />
-                            </div>
-                        </div>
-                        <ReferenceForm
-                            class="mt-2 w-90"
-                            @add="addEntityReference"
-                        />
-                    </div>
-                    <hr
-                        v-if="state.hasEntityReferences && state.hasAttributeReferences"
-                        class="w-90"
-                    >
-                    <template v-if="state.hasAttributeReferences">
-                        <template
-                            v-for="(referenceGroup, key) in state.attributeReferences"
-                            :key="key"
-                        >
-                            <div
-                                v-if="referenceGroup.length > 0"
-                                class="reference-group"
-                            >
-                                <h5 class="mb-2 fw-medium">
-                                    <a
-                                        href="#"
-                                        class="text-decoration-none"
-                                        @click.prevent="showMetadataForReferenceGroup(referenceGroup)"
-                                    >
-                                        {{ translateConcept(key) }}
-                                    </a>
-                                </h5>
-                                <div class="list-group w-90">
-                                    <div
-                                        v-for="(reference, i) in referenceGroup"
-                                        :key="i"
-                                        class="list-group-item pt-0"
-                                    >
-                                        <header class="text-end">
-                                            <span class="text-muted fw-light small">
-                                                {{ date(reference.updated_at) }}
-                                            </span>
-                                        </header>
-                                        <Quotation :value="reference" />
-                                    </div>
-                                </div>
-                            </div>
-                        </template>
-                    </template>
-                </div>
+                <ReferenceTab />
             </div>
         </div>
     </div>
@@ -160,27 +85,10 @@
         useRoute,
     } from 'vue-router';
 
-    import useBibliographyStore from '@/bootstrap/stores/bibliography.js';
     import useEntityStore from '@/bootstrap/stores/entity.js';
     import useSystemStore from '@/bootstrap/stores/system.js';
     import router from '%router';
 
-    import {
-        translateConcept,
-    } from '@/helpers/helpers.js';
-
-    import {
-        formatAuthors,
-    } from '@/helpers/bibliography.js';
-
-    import {
-        date,
-    } from '@/helpers/filters.js';
-
-    import {
-        canShowReferenceModal,
-        showLiteratureInfo,
-    } from '@/helpers/modal.js';
     import {
         subscribeNotifications,
         subscribeSystemChannel,
@@ -199,22 +107,16 @@
         handleNotifications,
     } from '@/handlers/notification.js';
 
-    import { useToast } from '@/plugins/toast.js';
-
-    import Quotation from '@/components/bibliography/Quotation.vue';
-    import ReferenceForm from '@/components/bibliography/ReferenceForm.vue';
     import useWebSocketConnectionToast from '@/composables/websocket-connection-toast.js';
+    import ReferenceTab from './bibliography/ReferenceTab.vue';
 
     export default {
         components: {
-            Quotation,
-            ReferenceForm,
+            ReferenceTab,
         },
         setup(props, context) {
             const { t } = useI18n();
             const currentRoute = useRoute();
-            const toast = useToast();
-            const bibliographyStore = useBibliographyStore();
             const entityStore = useEntityStore();
             const systemStore = useSystemStore();
             useWebSocketConnectionToast();
@@ -229,46 +131,15 @@
                     append: true,
                 });
             };
-            const isTab = id => {
-                return state.tab == id;
-            };
-            const addEntityReference = data => {
-                entityStore.addReference(state.entity.id, null, null, data);
-            };
-            const showMetadataForReferenceGroup = referenceGroup => {
-                if(!referenceGroup) return;
-                if(!state.entity) return;
-                const aid = referenceGroup[0].attribute_id;
-
-                const canOpen = canShowReferenceModal(aid);
-                if(canOpen) {
-                    router.push({
-                        append: true,
-                        name: 'entityrefs',
-                        query: currentRoute.query,
-                        params: {
-                            aid: aid,
-                        },
-                    });
-                } else {
-                    const msg = t('main.entity.references.toasts.cannot_edit_metadata.msg');
-                    toast.$toast(msg, '', {
-                        duration: 2500,
-                        autohide: true,
-                        channel: 'warning',
-                        icon: true,
-                        simple: true,
-                    });
-                }
-            };
-            const openLiteratureInfo = reference => {
-                showLiteratureInfo(reference.bibliography.id);
-            };
-
+           
             // DATA
             const state = reactive({
                 tab: computed(_ => systemStore.mainView.tab),
                 tabComponent: computed(_ => {
+                    if(state.tab === 'references') {
+                        return ReferenceTab;
+                    }
+
                     const plugin = state.tabPlugins.find(p => p.key == state.tab);
                     if(!!plugin) {
                         return plugin.componentTag;
@@ -278,38 +149,6 @@
                 }),
                 concepts: computed(_ => systemStore.concepts),
                 entity: computed(_ => entityStore.selectedEntity),
-                hasEntityReferences: computed(_ => {
-                    const isNotSet = !state.entity.references;
-                    if(isNotSet) return false;
-
-                    const isEmpty = !Object.keys(state.entity.references).length > 0;
-                    if(isEmpty) return false;
-                    return state.entity.references.on_entity?.length > 0;
-                }),
-                hasAttributeReferences: computed(_ => {
-                    const isNotSet = !state.entity.references;
-                    if(isNotSet) return false;
-                    const {
-                        on_entity,
-                        ...refs
-                    } = state.entity.references;
-                    const isEmpty = !Object.keys(refs).length > 0;
-                    if(isEmpty) return false;
-                    return Object.values(refs).some(v => v.length > 0);
-                }),
-                hasReferences: computed(_ => state.hasEntityReferences || state.hasAttributeReferences),
-                entityReferences: computed(_ => state.hasEntityReferences ? state.entity.references.on_entity : []),
-                attributeReferences: computed(_ => {
-                    if(state.hasAttributeReferences) {
-                        const {
-                            on_entity,
-                            ...refs
-                        } = state.entity.references;
-                        return refs;
-                    } else {
-                        return {};
-                    }
-                }),
                 entityTypes: computed(_ => entityStore.entityTypes),
                 columnPref: computed(_ => systemStore.getPreference('prefs.columns')),
                 isDetailLoaded: computed(_ => state.entity?.id > 0),
@@ -347,15 +186,8 @@
             return {
                 t,
                 // HELPERS
-                translateConcept,
-                formatAuthors,
-                date,
                 // LOCAL
                 setTab,
-                isTab,
-                addEntityReference,
-                showMetadataForReferenceGroup,
-                openLiteratureInfo,
                 // STATE
                 state,
             };
